% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_pb}
\alias{apply_pb}
\title{Plots validity metrics}
\usage{
apply_pb(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{an filtR object}

\item{MARGIN}{a metric}

\item{FUN}{a caption for each plot}

\item{...}{TBD}
}
\value{
TBD
}
\description{
Plots validity metrics
}
