% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\docType{data}
\name{dane}
\alias{dane}
\title{Sztuczny zbior danych}
\format{
Ramka denych, ktora ma 3 kolumny i 2289 wierszy

\code{data} - data z kazdego dnia od 2014-11-01 do 2021-02-05,

\code{cena} - cena z rozkladu normalnego o srednej 1000 i odchylenie standardowe 100,

\code{ilosc} - zbudowana losowo sposrod liczb od 1 do 2289 - losowanie liczb ze zwracaniem.
}
\usage{
dane
}
\description{
Zbior danych sztucznie wygenerowany.
}
\keyword{datasets}
