% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtR_function.R
\name{filtR}
\alias{filtR}
\title{Filters OTUs Generated by PCR or Sequencing Error from Count Tables}
\usage{
filtR(count_file, rho_CO = 0.7, clr_CO = 5, plot = FALSE)
}
\arguments{
\item{count_file}{Input your saved count table file.}

\item{rho_CO}{Input your rho value cutoff. Range of 0 to 1. Higher value results in a tighter filter with fewer OTUs being discarded. Defaults to 0.7.}

\item{clr_CO}{Input your centered log ratio difference cutoff. Higher value results in a tighter filter with fewer OTUs being discarded. Defaults to 5.}

\item{plot}{Set to TRUE to create scatter plots of the CLRs of the filtered OTU plotted against the CLRs of the OTU from which they are derived. The black, solid line is the ideal slope of the line for association (slope of 1) and the red dashed line is the slope of the linear model of the data. Defaults to FALSE.}
}
\value{
Count table with error-derived OTUs removed.
}
\description{
Using propr and ALDEx2 functions, filtR removes OTUs generated by PCR and sequencing error from count tables. The input for this package is a count table with the columns as samples and the rows as OTUs/ASVs.
}
\examples{
function(count_file='SRP058027_taxonomy_abundances_SSU_v4.1.tsv', rho_CO=0.7, clr_CO=5, plot=TRUE)
}
