% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filtr1}
\alias{filtr1}
\title{Filtrowanie z limitem dolnym i gornym}
\usage{
filtr1(dane, kolumna, limit_dolny, limit_gorny, flaga = FALSE)
}
\arguments{
\item{dane}{Zbior danych uzytkownika jako data frame. Funkcje te dzialaja na ramce danych zawierającej 3 kolumny:
\itemize{
\item data (rok-miesiąc-dzień),
\item cena,
\item ilosc.
}}

\item{kolumna}{Kolumna ktora zostanie wykorzystana do dokonywania selekcji}

\item{limit_dolny}{Najmniejsza akceptowalna wartosc z danej kolumny}

\item{limit_gorny}{Najwieksza akceptowalna wartosc z danej kolumny}

\item{flaga}{Zmienna logiczna, domyslnie \code{FALSE}.}
}
\value{
Funkcja zwraca:
\itemize{
\item dla \code{FALSE} zawezona ramke (tylko rekordy spelniajace warunki filtru),
\item dla \code{TRUE} z pelna ramke z dodatkowa kolumna FLAGA (1, gdy wiersz/rekord spelnia krytrium filtru, 0 gdy nie spelnia).
}
}
\description{
Pierwsza funkcja dokonuje selekcji tych rekordow, dla ktorych wskazana kolumna (np. cena)
jest między limit_dolny a limit_górny.
}
\examples{
filtr1(dane,"cena", 600, 850)
filtr1(dane,"ilosc", 800, 1750, TRUE)
}
