% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filtr2}
\alias{filtr2}
\title{Filtrowanie z kwantylem dolnym i gornym}
\usage{
filtr2(dane, kolumna, kwantyl_dolny, kwantyl_gorny, flaga = FALSE)
}
\arguments{
\item{dane}{Zbior danych uzytkownika jako data frame. Funkcje te dzialaja na ramce danych zawierającej 3 kolumny:
\itemize{
\item data (rok-miesiąc-dzień),
\item cena,
\item ilosc.
}}

\item{kolumna}{Kolumna ktora zostanie wykorzystana do dokonywania selekcji}

\item{kwantyl_dolny}{Limit dolny}

\item{kwantyl_gorny}{Limit gorny}

\item{flaga}{Zmienna logiczna, domyslnie \code{FALSE}.}
}
\value{
Funkcja zwraca:
\itemize{
\item dla \code{FALSE} zawezona ramke (tylko rekordy spelniajace warunki filtru),
\item dla \code{TRUE} z pelna ramke z dodatkowa kolumna FLAGA (1, gdy wiersz/rekord spelnia krytrium filtru, 0 gdy nie spelnia).
}
}
\description{
Oblicza kwantyle rozkladu wartosci ze wskazanej kolumny
(odpowiednio rzedu kwantyl dolny i kwantyl gorny) i dokonuje selekcji
tych wartosci z danej kolumny, ktore sie mieszcza pomiedzy
}
\examples{
filtr2(dane,"cena", 0.15, 0.35)
filtr2(dane,"ilosc", 0.25, 0.75, TRUE)
}
