% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_comb}
\alias{get_comb}
\title{Get all possible filter combinations}
\usage{
get_comb(effvar, efffac, filtervars, df, sample)
}
\value{
a df with all combinations
}
\description{
Get all possible filter combinations
}
