% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_filter}
\alias{get_filter}
\title{Checking the validity of dataset}
\usage{
get_filter(outcome, treatment, filtervars = NULL, data, control = FALSE)
}
\arguments{
\item{filtervars}{numeric or factor, vector of column names}

\item{data}{numeric or factor, a dataset}

\item{out}{numeric, the column name of the}

\item{treat}{numeric or factor, the column name}

\item{sample}{boolean, if TRUE sample from observations}

\item{plot}{boolean, if TRUE call plot.valid}

\item{smooth}{boolean, if TRUE call smooth.valid}
}
\value{
a filtR object including estimations of effect size an power for all possible filtervariable x value combination
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
data <- data.frame(a = c(1:200), b = c(201:400), c = factor(rep(1:2, 100)), d = c(201:400))
get_filter(outcome = "a", treatment = "b", data = data)
}
