% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_ids}
\alias{get_ids}
\title{Subset dataset given a combinatiuon of filter variables}
\usage{
get_ids(grid, data, cols)
}
\arguments{
\item{data}{numeric or factor, a dataset}

\item{cols}{A vector of class "character"}
}
\value{
A vector of class "integer", which includes the rows that are within the filter thresholds
}
\description{
Subset dataset given a combinatiuon of filter variables
}
