% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_plot}
\alias{get_plot}
\title{Plots validity metrics}
\usage{
get_plot(x, caption = c("Effect Size vs. Filter", "Power vs. Filter"),
  main = "", ...)
}
\arguments{
\item{x}{an filtR object}

\item{caption}{a caption for each plot}

\item{main}{a title for each plot}

\item{...}{Arguments to be passed to methods}
}
\value{
a plot of effect size and power for all combinations
}
\description{
Plots validity metrics
}
