% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{get_point}
\alias{get_point}
\title{Checking the validity of dataset}
\usage{
get_point(outcome, treatment, exp, method = "effect", data)
}
\arguments{
\item{outcome}{numeric, the column name of the}

\item{treatment}{numeric or factor, the column name}

\item{exp}{numeric or factor, vector of column names}

\item{method}{character, ...}

\item{data}{numeric or factor, a dataset}
}
\value{
tbd
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
}
