% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_results}
\alias{get_results}
\title{Compute model output for each possible combination}
\usage{
get_results(combination, data, formula, treatment)
}
\arguments{
\item{combination}{TBD}

\item{data}{TBD}

\item{formula}{TBD}
}
\value{
An object of class "data.frame", which includes the model output for the given combination of filter variables
}
\description{
Compute model output for each possible combination
}
