% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_subset}
\alias{get_subset}
\title{Subsets original dataset and calculates validity metrics}
\usage{
get_subset(x, effvar, efffac, data)
}
\arguments{
\item{x}{combination passed}

\item{effvar}{numeric, the column name of the}

\item{efffac}{numeric or factor, the column name}

\item{data}{numeric or factor, a dataset}
}
\value{
a vector with validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
