% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.filtr}
\alias{plot.filtr}
\title{Plots validity metrics}
\usage{
\method{plot}{filtr}(x, metric, caption, main, ...)
}
\arguments{
\item{x}{an filtR object}

\item{metric}{a metric}

\item{caption}{a caption for each plot}

\item{main}{a title for each plot}

\item{...}{Arguments to be passed to methods}
}
\value{
a plot of effect size and power for all combinations
}
\description{
Plots validity metrics
}
