% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{valid.subset}
\alias{valid.subset}
\title{Subsets original dataset and calculates validity metrics}
\usage{
valid.subset(x, effvar, efffac, dat)
}
\arguments{
\item{x}{input from wrapper}

\item{effvar}{the name of the}

\item{efffac}{bla}

\item{dat}{bla}
}
\value{
a vector with validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
