% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{valid_subset}
\alias{valid_subset}
\title{Subsets original dataset and calculates validity metrics}
\usage{
valid_subset(x, effvar, efffac, dat)
}
\arguments{
\item{effvar}{numeric, the column name of the}

\item{efffac}{numeric or factor, the column name}
}
\value{
a vector with validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
