% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\docType{data}
\name{dane}
\alias{dane}
\title{Sztuczny zbior danych}
\format{
Ramka denych, ktora ma 3 kolumny i 2289 wierszy

\code{data} - Data w formacie rok-miesiac-dzien

\code{cena} - Ceny (PLN)

\code{ilosc} - Ilosc (szt)
}
\usage{
dane
}
\description{
Zbior danych sztucznie wygenerowanych
}
\keyword{datasets}
