% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtR_function.R
\name{filtR}
\alias{filtR}
\title{Filters OTUs Generated by PCR or Sequencing Error from Count Tables}
\usage{
filtR(count_file, rho_CO = 0.7, clr_CO = 5)
}
\arguments{
\item{count_file}{Input your saved count file.}

\item{rho_CO}{Input your rho value cutoff. Range of 0 to 1. Higher value results in a tighter filter with less OTUs being discarded. Defaults to 0.7.}

\item{clr_CO}{Input your centered log ratio difference cutoff. Higher value results in a tighter filter with less OTUs being discarded Defaults to 5.}
}
\value{
Count table with error-derived OTUs removed.
}
\description{
Using propr and ALDEx2 functions, filtR removes OTUs generated by PCR and sequencing error from count tables. There are two filtering parameters. One is rho, which is a measure of beta-association between two OTUs. Each individual OTU, for the most part, should lack association with other OTUs because they are independent entities. The other filtering parameter is difference in centered log ratio. If an OTU is derived from an error in PCR or sequencing, it is expected to have a much lower abundance than the OTU it associates with. The input for this package is a count table with the columns as samples and the rows as OTUs/ASVs. The output is a count table with the error-derived OTUs removed.
}
\examples{
function(count_file='SRP058027_taxonomy_abundances_SSU_v4.1.tsv', rho_CO=0.7, clr_CO=5)
}
