% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_comb}
\alias{get_comb}
\title{Get all possible filter combinations}
\usage{
get_comb(effvar, efffac, filtervars, data, sample)
}
\arguments{
\item{effvar}{numeric, the column name of the}

\item{efffac}{numeric or factor, the column name}

\item{filtervars}{numeric or factor, vector of column names}

\item{data}{numeric or factor, a dataset}

\item{sample}{boolean, if TRUE sample from observations}
}
\value{
a data with all combinations
}
\description{
Get all possible filter combinations
}
