% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_comb}
\alias{get_comb}
\title{Get all possible filter combinations}
\usage{
get_comb(outcome, treatment, data)
}
\arguments{
\item{outcome}{numeric, the column name of the}

\item{treatment}{numeric or factor, the column name}

\item{data}{numeric or factor, a dataset}
}
\value{
An object of class "data.frame", which includes all unique possible combinations of filter variables
}
\description{
Get all possible filter combinations
}
