% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{get_filter}
\alias{get_filter}
\title{Checking the validity of dataset}
\usage{
get_filter(outcome, treatment, data)
}
\arguments{
\item{outcome}{numeric, the column name of the}

\item{treatment}{numeric or factor, the column name}

\item{data}{numeric or factor, a dataset}
}
\value{
a filtR object including estimations of effect size an power for all possible filtervariable x value combination
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
data <- data.frame(outcome = c(1:200), treatment = sample(rep(c(1, 2), 100)), filt = rnorm(200), filt2 = rnorm(200))
get_filter(outcome = "outcome", treatment = "treatment", data = data)
}
