% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{plot.valid}
\alias{plot.valid}
\title{Plots validity metrics}
\usage{
\method{plot}{valid}(obj, caption = c("Effect Size vs. Filter",
  "Power vs. Filter"), main = "")
}
\arguments{
\item{obj}{an filtR object}

\item{caption}{a caption for each plot}

\item{main}{a title for each plot}
}
\value{
a plot of effect size and power for all combinations
}
\description{
Plots validity metrics
}
