% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{plot.valid}
\alias{plot.valid}
\title{Subsets original dataset and calculates validity metrics}
\usage{
\method{plot}{valid}(x, caption = c("Effect Size vs. Filter",
  "Power vs. Filter"), main = "")
}
\arguments{
\item{x}{an filtR object}
}
\value{
a plot of validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
