% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subset_fun}
\alias{subset_fun}
\title{Subsets original dataset and calculates validity metrics}
\usage{
subset_fun(x)
}
\arguments{
\item{x}{input from wrapper}
}
\value{
a tibble with validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
