% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subset_fun}
\alias{subset_fun}
\title{Subsets original dataset and calculates validity metrics}
\usage{
subset_fun(x, target1, target2 = NULL, groupvar = NULL, dat)
}
\arguments{
\item{x}{input from wrapper}

\item{target1}{the name of the}

\item{target2}{bla}

\item{groupvar}{bla}

\item{dat}{bla}
}
\value{
a vector with validity criteria
}
\description{
Subsets original dataset and calculates validity metrics
}
