% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{valid}
\alias{valid}
\title{Checking the validity of dataset}
\usage{
valid(target1, target2 = NULL, groupvar = NULL, filtervar = NULL, df,
  plot = FALSE, smooth = FALSE)
}
\arguments{
\item{target1}{the name of the}

\item{target2}{a}

\item{groupvar}{a treatment variable}

\item{plot}{call plot.valid}

\item{smooth}{call smooth.valid}

\item{dat}{a dataset}
}
\value{
a filtR object including estimations of effect size an power for each combination
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
dat <- data.frame(a = c(1:200), b = c(201:400), c = factor(rep(1:2, 100)), d = c(201:400))
target1 <- "a"
target2 <- "b"
valid(target1 = target1, target2 = target2, dat = dat)

}
