% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{valid}
\alias{valid}
\title{Checking the validity of dataset}
\usage{
valid(target1, target2 = NULL, groupvar = NULL, dat)
}
\arguments{
\item{target1}{the name of the}

\item{target2}{bla}

\item{groupvar}{bla}

\item{dat}{bla}
}
\value{
a data.frame including
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
dat <- data.frame(a = c(1:200), b = c(201:400), c = factor(rep(1:2, 100)), d = c(201:400))
target1 <- "a"
target2 <- "b"
valid(target1 = target1, target2 = target2, dat = dat)

}
