% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{valid}
\alias{valid}
\title{Checking the validity of dataset}
\usage{
valid(effvar, efffac, filtervars = NULL, df, sample = FALSE,
  plot = FALSE, smooth = FALSE)
}
\arguments{
\item{effvar}{numeric, the column name of the}

\item{efffac}{numeric or factor, the column name}

\item{filtervars}{numeric or factor, vector of column names}

\item{df}{numeric or factor, a dataset}

\item{sample}{boolean, if TRUE sample from observations}

\item{plot}{boolean, if TRUE call plot.valid}

\item{smooth}{boolean, if TRUE call smooth.valid}
}
\value{
a filtR object including estimations of effect size an power for all possible filtervariable x value combination
}
\description{
Checking the validity of dataset
}
\examples{
library(filtR)
data <- data.frame(a = c(1:200), b = c(201:400), c = factor(rep(1:2, 100)), d = c(201:400))
valid(effvar = "a", efffac = "b", df = data)
}
